feat: delete folder data
Updated the delete folder endpoint to update book_data.json to reflect the deleted folder.
Added BD_delete_folder method to delete folders from book_data.json and move their books to the "misc" folder if requested.
Added a data provider parameter to BD_move_book that allows it to function when called by another function that manipulates book_data.json, similar to delete_book (it is called multiple times by delete_folder to extract books from the folder's array before it is removed).
Tested rename folder data to ensure functionality

+++ Assets/Scripts/API_segments/Folder_methods.py
@@ -1,6 +1,6 @@
 import os
 import shutil
-from .Book_data_endpoint_methods import BD_rename_folder
+from .Book_data_endpoint_methods import BD_rename_folder, BD_delete_folder
 
 
 def list_folders(mainDir):
@@ -46,14 +46,20 @@ def Delete_folder(folder_name, delete_content, mainDir):
                     else:
                         os.rename(start, destination)
 
-            shutil.rmtree(target)
+            # Update the book data
+            result = BD_delete_folder(folder_name, delete_content)
+            if result == "200":
+                shutil.rmtree(target)
+                if (notifyChange):
+                    return "428"
+                return result
+            else:
+                raise Exception(result)
+
         except Exception as e:
             return "500: " + str(e)
     else:
         return "404"
-    if (notifyChange):
-        return "428"
-    return "200"
 
 
 def Rename_folder(folder_name, new_name, mainDir):
